refactor schema file to use axios for query external api instead of fetch
