+++ package.json
@@ -6,6 +6,7 @@
     "apollo-datasource-rest": "^0.3.2",
     "apollo-server": "^2.4.6",
     "apollo-server-express": "^2.4.6",
+    "axios": "^0.18.0",
     "dotenv": "^6.2.0",
     "express": "^4.16.4",
     "express-graphql": "^0.7.1",
